pyproject.toml
[tool.poetry]
name = "binance-orders"
version = "0.1.0"
description = ""
authors = ["Dm1triiSmirnov <smirnov0719@gmail.com>"]
readme = "README.md"
packages = [{include = "binance_orders"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.1"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
python-binance = "^1.0.17"
binance = "^0.3"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
